requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.10
Jinja2==3.1.5
Werkzeug==3.1.3
MarkupSafe==3.0.2
gunicorn==23.0.0
